requirements.txt
alembic==1.7.5
anyio==3.3.4
asciinema==2.1.0
asgiref==3.4.1
bcrypt==3.2.0
CacheControl==0.12.10
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
cryptography==35.0.0
DateTime==4.3
distlib==0.3.4
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
filelock==3.4.2
Flask==2.0.2
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httptools==0.2.0
hypercorn==0.12.0
hyperframe==6.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
keyring==21.8.0
lockfile==0.12.2
Mako==1.1.6
MarkupSafe==2.0.1
mongoengine==0.23.1
motor==2.5.1
msgpack==1.0.3
orjson==3.6.4
packaging==20.9
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.8.2
platformdirs==2.4.1
poetry==1.1.12
poetry-core==1.0.7
priority==2.0.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyJWT==1.7.1
pylev==1.4.0
pymongo==3.12.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==5.4.1
requests==2.26.0
requests-toolbelt==0.9.1
rsa==4.7.2
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.5
starlette==0.16.0
tenacity==8.0.1
toml==0.10.2
tomlkit==0.8.0
tqdm==4.62.3
typing==3.7.4.3
typing-extensions==3.7.4.3
ujson==4.2.0
urllib3==1.26.7
uvicorn==0.15.0
virtualenv==20.13.0
watchgod==0.7
webencodings==0.5.1
websockets==10.0
Werkzeug==2.0.2
wsproto==1.0.0
zope.interface==5.4.0
